Added confirm modal windows when deleting steps in edit and new post views

+++ answerharbor_app/static/js/vue/edit_post_app.js
@@ -97,7 +97,7 @@ var app = new Vue({
             this.stepGroup.addNewStep();
         },
         deleteStep: function(step) {
-            if(this.stepGroup.steps.length > 1){
+            if(this.stepGroup.steps.length > 1 && confirm('Are you sure you want to delete this step?')){
                 var stepInputNameID = step.inputNameID;
                 this.stepGroup.deleteStep(step);
             }
